backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1

# Resume parsing libraries (we'll test these)
pyresparser==1.0.6
spacy==3.7.2
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1
pdfminer.six==20221105



gunicorn==21.2.0
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
requests
password-validator
bcrypt==3.2.2


# Core dependencies for resume parsing
pdfplumber==0.10.3
python-docx==1.1.0

# OCR dependencies
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Additional image processing (optional but recommended)
opencv-python==4.8.1.78

# For better text preprocessing
regex==2023.10.3

https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl

geonamescache
spacy 
en_core_web_sm
en_core_web_md
Spire.Doc
docling
locationtagger
markitdown
pypandoc
olefile
textract